Reviewed entire logic for the WhatsNewLocalDataProvider
